Refactor string formatting to use Rust's new formatting syntax
- Updated various instances of `format!` to use the new `{variable}` syntax for improved readability and consistency across the codebase.
- Modified performance metrics reporting in `performance_metrics.rs` for better clarity.
- Enhanced error and log messages in `profile_manager.rs`, `protocol.rs`, and `telnet_negotiation.rs` to utilize the new formatting style.
- Cleaned up terminal and UI related files to adopt the new formatting approach, ensuring a uniform style throughout the application.
- Removed unnecessary `Default` implementations and re-exported items in `protocol_common/mod.rs` for better organization.
- Improved test cases in `test_field_detection.rs` to reflect the new formatting style.

+++ src/terminal_display.rs
@@ -91,9 +91,9 @@ impl TN5250RApp {
             }
 
             // Draw cursor if it's beyond the text content
-            if cursor_pos.0 as usize > lines.len() ||
-                (cursor_pos.0 as usize <= lines.len() &&
-                 cursor_pos.1 as usize > lines.get(cursor_pos.0 - 1).map_or(0, |l| l.len())) {
+            if cursor_pos.0 > lines.len() ||
+                (cursor_pos.0 <= lines.len() &&
+                 cursor_pos.1 > lines.get(cursor_pos.0 - 1).map_or(0, |l| l.len())) {
 
                 let cursor_char_pos = rect.min + egui::vec2(
                     (cursor_pos.1 - 1) as f32 * char_width,
